requirements.txt
numpy>=1.26
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
langchain>=0.3
langchain-community
pypdf
load-dotenv
openai
tiktoken